requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0
websockets==12.0
python-jose==3.3.0
python-multipart==0.0.9
bcrypt==4.1.2
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic-settings==2.2.1